Finish the heroes routes

+++ src/app/heroes/pages/add-hero/add-hero.component.ts
@@ -4,6 +4,8 @@ import { HeroesService } from '../../services/heroes.service';
 import { ActivatedRoute, Router } from '@angular/router';
 import { switchMap } from 'rxjs';
 import { MatSnackBar } from '@angular/material/snack-bar';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmComponent } from '../../components/confirm/confirm.component';
 
 @Component({
   selector: 'app-add-hero',
@@ -43,7 +45,8 @@ export class AddHeroComponent implements OnInit {
   constructor( private heroesService: HeroesService,
                private activatedRoute: ActivatedRoute,
                private router: Router,
-               private snackBar:MatSnackBar ) { }
+               private snackBar:MatSnackBar,
+               public dialog: MatDialog ) { }
 
   ngOnInit(): void {
 
@@ -86,10 +89,23 @@ export class AddHeroComponent implements OnInit {
   }
 
   deleteHero() {
-    this.heroesService.deleteHero( this.hero.id! )
-    .subscribe( resp => {
-      this.router.navigate(['/heroes'])
-    })
+    const dialog = this.dialog.open( ConfirmComponent, {
+      width: '30rem',
+      data: this.hero
+    });
+
+    dialog.afterClosed().subscribe(
+      ( result ) => {
+
+        if( result ) {
+
+           this.heroesService.deleteHero( this.hero.id! )
+           .subscribe( resp => {
+             this.router.navigate(['/heroes']);
+           })
+        }
+      })
+
   }
 
   showSnackbar( message: string ) {
